requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
MetaTrader5>=5.0.45
PyJWT>=2.8.0
